requirements.txt
faker
pytest
kafka-python
loguru
pyspark==3.5.0
duckdb